+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+- Enum and interface naming has been changed
+- Single export file for generated files
+- From now on global interfaces are called initializers
+
 ## 0.1.1
 - README correction